feat: Add script to extract single-cell barcode from nano-CT sequencing
This commit adds a new helper script, `get_scCT_barcode_from_nano_CT_sequencing.py`, which extracts the single-cell barcode from nano-CT sequencing data. The script takes an input R2 fastq file from nano-CT sequencing and writes the 16bp single-cell barcode to an output file. It also includes an option to specify the number of threads to use for bgzip compression.

The purpose of this script is to facilitate the analysis of single-modality scCUT&Tag data that was sequenced using the nano-CT sequencing setup. The script extracts the single-cell barcode from the sequencing data, which can be used for downstream analysis.

.